Fix img convert

+++ create-image.sh
@@ -2,17 +2,21 @@
 
 set -ex
 
-# if HDA_SIZE starts with "+", add to computed size.
-if [[ $HDA_SIZE == +* ]]; then
-  BASE_SIZE=$(du -s "$CHROOT" | awk "{print int(\$1)}")
-  qemu-img create -f raw "$HDA_RAW" "${BASE_SIZE}k"
-  qemu-img resize -f raw "$HDA_RAW" "$HDA_SIZE"
-else
-  qemu-img create "$HDA_RAW" "$HDA_SIZE"
+if [ -e "$CHROOT" ] && [ ! -e "$HDA_RAW" ]; then
+  # if HDA_SIZE starts with "+", add to computed size.
+  if [[ $HDA_SIZE == +* ]]; then
+    BASE_SIZE=$(du -s "$CHROOT" | awk "{print int(\$1)}")
+    qemu-img create -f raw "$HDA_RAW" "${BASE_SIZE}k"
+    qemu-img resize -f raw "$HDA_RAW" "$HDA_SIZE"
+  else
+    qemu-img create "$HDA_RAW" "$HDA_SIZE"
+  fi
+
+  mkfs.ext2 -d "$CHROOT/" "$HDA_RAW"
+  [ ! -w "$CHROOT" ] || rm -rf "$CHROOT"
 fi
 
-mkfs.ext2 -d "$CHROOT/" "$HDA_RAW"
-[ -w "$CHROOT" ] && rm -rf "$CHROOT"
-
-qemu-img convert -f raw -O qcow2 "$HDA_RAW" "$HDA_QCOW2"
-[ -w "$HDA_RAW" ] && rm "$HDA_RAW"
+if [ -e "$HDA_RAW" ] && [ ! -e "$HDA_QCOW2" ]; then
+  qemu-img convert -f raw -O qcow2 "$HDA_RAW" "$HDA_QCOW2"
+  [ ! -w "$HDA_RAW" ] || rm "$HDA_RAW"
+fi
